CG-0MP0B51RK002C0EV: extract MainStreetOverlayManager game-over overlay
